fall2024-thesis/src/impl_rpi/requirements.txt
cbor==1.0.0
contourpy==1.3.0
coppeliasim_zmqremoteapi_client==2.0.4
cycler==0.12.1
fastdtw==0.3.4
fonttools==4.53.1
kiwisolver==1.4.7
loguru==0.7.2
matplotlib==3.9.2
msgpack==1.0.8
numpy==2.1.1
packaging==24.1
paho-mqtt==2.1.0
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyzmq==26.2.0
scipy==1.14.1
six==1.16.0
smbus2==0.4.3
